requirements.txt
auto_mix_prep==0.2.0
bitstruct==8.10.0
cycler==0.11.0
matplotlib==3.5.1
numpy==1.18.1
opencv_python==4.5.5.62
pandas==1.4.2
Pillow==9.1.0
PyQt5==5.15.6
scipy==1.8.0
Shapely==1.8.1.post1
SPAR==0.0.7
